connectivity: extract the sniffer logic into a separate package
Let's extract this logic into a separate package, so that it can be
easily reused by other connectivity tests in addition to the encryption
ones. At this stage, no additional changes are introduced except for
exporting the necessary types/methods and adapting the call sites.
